+++ src/routes/auth0pattern-web.php
@@ -1,14 +1,16 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use Auth0\Laravel\Controllers\LogoutController;
+use Auth0\Laravel\Controllers\CallbackController;
+// use Auth0\Laravel\Controllers\LoginController;               // Disabled Auth0 login controller...
+use FaithFM\Auth0Pattern\Http\Controllers\LoginController;      // ...replaced with our own version - which captures the 'previous' URL and saves it as the 'intended' URL
 
-// Auth0-related routes.   (Replaces Laravel's default auth routes - normally added with a "Auth::routes();" statement.)
-// Note: auth0/login v7 includes their own routes (supercedes our old Auth0PatternController ones... although we do lose our /profile route)
-Route::middleware(['web_group'])->group(function () {
-    Route::get('/login',                        \Auth0\Laravel\Http\Controller\Stateful\Login::class)->name('login');
-    Route::match(['get', 'post'], '/logout',    \Auth0\Laravel\Http\Controller\Stateful\Logout::class)->name('logout');
-    Route::get('/auth0/callback',               \Auth0\Laravel\Http\Controller\Stateful\Callback::class)->name('auth0.callback');
+// Note: we had to disable Auth0's automatic route registration in the config/auth0.php file (ie: 'registerAuthenticationRoutes' => false)...
+// ... since we are using our own 'web_group' group... and Auth0 is hard-coded to use the 'web' middleware group.
 
+Route::group(['middleware' => 'web_group'], static function (): void {
+  Route::get('/login', LoginController::class)->name('login');                          // Remember '/login' uses our own special controller
+  Route::match(['get', 'post'], '/logout', LogoutController::class)->name('logout');
+  Route::get('/callback', CallbackController::class)->name('callback');
 });
-
-