+++ templates/app/Models/User.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is cloned / force-published from the "laravel-auth0-pattern" composer package.
  *    WARNING: Local modifications will be overwritten when the package is updated.
@@ -7,20 +8,21 @@
 
 namespace App\Models;
 
-use Auth0\Laravel\Contract\Model\Stateful\User as StatefulUser;
 use Illuminate\Auth\Authenticatable;
-use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableUser;
+use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableContract;
+use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Foundation\Auth\Access\Authorizable;
 use Illuminate\Notifications\Notifiable;
 use OwenIt\Auditing\Contracts\Auditable;
 
-// Note: auth0/login v7's new methodology elimintates the need for our old Auth0PatternUserModelTrait.
-//    Apart from the "StatefulUser" interface implemented here, most Auth0-related functionality is now covered in our Auth0PatternUserRepository.php class.
+// Model signature is based on the Auth0 example: https://github.com/auth0/laravel-auth0/blob/main/docs/Eloquent.md#creating-a-user-model
 
-class User extends \Illuminate\Database\Eloquent\Model implements StatefulUser, AuthenticatableUser, Auditable
+class User extends Model implements AuthenticatableContract, AuthorizableContract, Auditable
 {
     use \OwenIt\Auditing\Auditable;
-    use HasFactory, Notifiable, Authenticatable;
+    use HasFactory, Notifiable, Authenticatable, Authorizable;
 
     /**
      * The attributes that are mass assignable.